# clients/android/contracts/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PhotoHouse native browsing consumer snapshot",
    "version": "1.0.0-fixture.1",
    "description": "Hand-authored consumer subset verified against pinned backend source and synthetic ASGI. Not a deployed server or complete API export."
  },
...
  "security": [
    {
      "nativeSession": []
    }
  ],
  "paths": {
...
    "/assets/{asset_id}/captions": {
      "get": {
        "operationId": "captions",
        "security": [
          {
            "nativeSession": []
          }
        ],
        "responses": {
          "400": {
            "description": "Malformed request or transport",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credential or object/library access denied; not a precise revocation diagnosis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Transport/boundary denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Refresh account state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 2048 bytes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Admission denied; honor Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Access unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Captions"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "description": "Positive int64 decimal; retain as a string."
            }
          },
          {
            "name": "library",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "nativeSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable 24-hour session; not JWT. No refresh endpoint."
      }
    },
    "schemas": {
...
      "Caption": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "Positive int64 decimal; retain as a string."
          },
          "text": {
            "type": "string"
          },
          "truncated": {
            "type": "boolean"
          },
          "user_edited": {
            "type": "boolean"
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "text",
          "truncated",
          "user_edited",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "Captions": {
        "type": "object",
        "properties": {
          "library_id": {
            "type": "string"
          },
          "asset_id": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "Positive int64 decimal; retain as a string."
          },
          "has_more": {
            "type": "boolean",
            "description": "True when rows are omitted by the 20-row or 524288-byte response limit; independent of per-item text truncation."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Caption"
            }
          }
        },
        "required": [
          "library_id",
          "asset_id",
          "has_more",
          "items"
        ],
        "additionalProperties": false,
        "description": "Complete encoded UTF-8 JSON response is at most 524288 bytes including envelope and escaping; ordered prefix of up to 20 whole rows. No caption pagination endpoint."
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "additionalProperties": false
      }
    }
  }
}